% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{input_filepath}
\alias{input_filepath}
\title{file io}
\arguments{
\item{input_filepath}{the path of the input PDF file. The default is set to
NULL. IF NULL, it  prompt the user to select the folder interactively.}
}
\description{
file io
}
\keyword{internal}
