% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{output_filepath}
\alias{output_filepath}
\title{file io}
\arguments{
\item{output_filepath}{the path of the output PDF file. The default is set to
NULL. IF NULL, it  prompt the user to select the folder interactively.}
}
\description{
file io
}
\keyword{internal}
