% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_pages.R
\name{remove_pages}
\alias{remove_pages}
\title{Remove selected pages from a file}
\usage{
remove_pages(rmpages, input_filepath = NULL, output_directory = "trim",
  output_filename = "trimmed_pdf")
}
\arguments{
\item{rmpages}{a vector of page numbers to be removed}

\item{input_filepath}{the path of the input PDF file.
The default is set to NULL. IF NULL, it  prompt the user to
select the folder interactively.}

\item{output_directory}{the name of the output directory}

\item{output_filename}{the name of the output file.}
}
\value{
this function returns a PDF document with the
remaining pages
}
\description{
If the toolkit Pdftk is available in the
system, it will be called to remove the given pages from
the seleted PDF files.

See the reference for detailed usage of \code{pdftk}.
}
\examples{
\dontrun{
# This command promts the user to select the file interactively.
# Remove page 2 and 3 from the selected file.
remove_pages(rmpages = c(3,6))
}
}
\references{
\url{https://www.pdflabs.com/tools/pdftk-the-pdf-toolkit/}
}
