% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fileIO.R
\name{return}
\alias{return}
\alias{fileIO}
\title{Return}
\usage{
fileIO(input_filepath, output_filepath, overwrite, system_command)
}
\value{
TRUE if the operation was succesfful. FALSE if the operation fails.
}
\description{
Return
}
\keyword{internal}
