% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rotate_pages.R
\name{rotate_pages}
\alias{rotate_pages}
\title{Rotate selected pages of a file to 90 degrees clockwise}
\usage{
rotate_pages(rotatepages, input_filepath = NULL, output_directory = NULL,
  output_filename = "rotated_pgs_pdf")
}
\arguments{
\item{rotatepages}{a vector of page numbers to be rotated  to 90 degrees clockwise}

\item{input_filepath}{the path of the input PDF file.
The default is set to NULL. IF NULL, it  prompt the user to
select the folder interactively.}

\item{output_directory}{the path of the output directory}

\item{output_filename}{the name of the output file.}
}
\value{
this function returns a PDF document with the
remaining pages
}
\description{
If the toolkit Pdftk is available in the
system, it will be called to rotate the given pages of
the seleted PDF files to 90 degrees clockwise.

See the reference for detailed usage of \code{pdftk}.
}
\examples{
\dontrun{
# This command promts the user to select the file interactively.
# Rotate page 2 and 6 to 90 degrees clockwise
rotate_pages(rotatepages = c(3,6))
}

\dontrun{
dir <- tempdir()
require(lattice)
for(i in 1:3) {
pdf(file.path(dir, paste("plot", i, ".pdf", sep = "")))
print(xyplot(iris[,1] ~ iris[,i], data = iris))
dev.off()
}
staple_pdf(input_directory = dir, output_directory = dir, output_filename = "Full_pdf")
input_path <- file.path(dir, paste("Full_pdf.pdf",  sep = ""))
rotate_pages(rotatepages = c(2,3), input_filepath = input_path, output_directory = dir)
}
}
\references{
\url{https://www.pdflabs.com/tools/pdftk-the-pdf-toolkit/}
}
\author{
Priyanga Dilini Talagala
}
