% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_from.R
\name{split_from}
\alias{split_from}
\title{Splits single input PDF document into two parts from a given point}
\usage{
split_from(pg_num, input_filepath = NULL, output_directory = NULL)
}
\arguments{
\item{pg_num}{a nonnegative integer. Split the pdf document into two from this page number.}

\item{input_filepath}{the path of the input PDF file.
The default is set to NULL. IF NULL, it  prompt the user to
select the folder interactively.}

\item{output_directory}{the path of the output directory}
}
\value{
this function splits a single input PDF document into
individual pages
}
\description{
If the toolkit Pdftk is available in the
system, it will be called to Split a single input PDF document
into two parts from a given point

See the reference for detailed usage of \code{pdftk}.
}
\examples{
\dontrun{
# Split the pdf from page 10
split_from(pg_num=10)
}

\dontrun{
dir <- tempdir()
require(lattice)
for(i in 1:4) {
pdf(file.path(dir, paste("plot", i, ".pdf", sep = "")))
print(xyplot(iris[,1] ~ iris[,i], data = iris))
dev.off()
}
staple_pdf(input_directory = dir, output_directory = dir)
input_path <- file.path(dir, paste("Full_pdf.pdf",  sep = ""))
split_from(pg_num=2, input_filepath = input_path ,output_directory = dir )
}
}
\references{
\url{https://www.pdflabs.com/tools/pdftk-the-pdf-toolkit/}
}
\author{
Priyanga Dilini Talagala
}
