% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/split_pdf.R
\name{split_pdf}
\alias{split_pdf}
\title{Splits a single input PDF document into individual pages.}
\usage{
split_pdf(input_filepath = NULL, output_directory = "split_folder")
}
\arguments{
\item{input_filepath}{the path of the input PDF file.
The default is set to NULL. IF NULL, it  prompt the user to
select the folder interactively.}

\item{output_directory}{the name of the output directory}
}
\value{
this function splits a single input PDF document into
individual pages
}
\description{
If the toolkit Pdftk is available in the
system, it will be called to Split a single input PDF document
into individual pages.

See the reference for detailed usage of \code{pdftk}.
}
\examples{
\dontrun{
split_pdf()
}
}
\references{
\url{https://www.pdflabs.com/tools/pdftk-the-pdf-toolkit/}
}
