% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staple_pages.R
\name{staple_pages}
\alias{staple_pages}
\title{Combine multiple PDF files}
\usage{
staple_pages(
  input_files,
  output_filepath = "output.pdf",
  start_pages = NA,
  end_pages = NA
)
}
\arguments{
\item{input_files}{Vector with paths of PDF files to be combined.}

\item{output_filepath}{PDF file path result of the combination.}

\item{start_pages}{Vector with the initial pages of each file. If \code{NA},
the default, will be considered the first page.}

\item{end_pages}{Vector with the final pages of each file. If \code{NA}, the
default, will be considered the last page.}
}
\value{
In the path informed in \code{output}, the PDF file resulting from the combination
of multiple files passed to \code{vec_output} will be saved.
}
\description{
Combine multiple PDF files by delimiting the sequences of pages in each file.
}
\examples{

\dontrun{
staple_pages(
   input_files =
      c(
        "file_1.pdf",
        "file_2.pdf",
      ),
   output_filepath = "output.pdf",
   start_pages = c(NA, NA),
   end_pages = c(NA, NA)
 )
}
}
