% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/staplr.R
\docType{package}
\name{staplr}
\alias{staplr}
\alias{staplr-package}
\title{staplr: A package containing a toolkit for PDF files}
\description{
This package provides function to manipulate PDF
files: merging multiple PDF files into one.
}
\references{
\url{https://www.pdflabs.com/tools/pdftk-the-pdf-toolkit/}
}
\seealso{
The core functions in this package: \code{\link{staple_pdf}}
}
\author{
Priyanga Dilini Talagala
}
