% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rename_files.R
\name{rename_files}
\alias{rename_files}
\title{Rename multiple files}
\usage{
rename_files(input_filepaths = NULL, new_names)
}
\arguments{
\item{input_filepaths}{the path of the input PDF files.
The default is set to NULL. IF NULL, it  prompt the user to
select the folder interactively.}

\item{new_names}{a vector of names for the output files.}
}
\value{
this function writes renamed files back to directory
}
\description{
Rename multiple files in a directory and
write renamed files back to directory
}
\examples{
\dontrun{
#if the directory contains 3 PDF files
rename_files(new_names = c("file 1", "file 2", "file 3"))
}
}
\references{
\url{https://www.pdflabs.com/tools/pdftk-the-pdf-toolkit/}
}
